requirements.txt
streamlit>=1.30.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy<2.0.0
pandas
pillow
opencv-python-headless
plotly
matplotlib
fpdf